+++ src/adapter/graphql/tests/tests/flows/test_gql_dataset_flow_processes.rs
@@ -300,6 +300,7 @@ async fn test_ingest_process_several_runs() {
                     pull_result: PullResult::Updated {
                         old_head: Some(odf::Multihash::from_digest_sha3_256(b"old-slice")),
                         new_head: odf::Multihash::from_digest_sha3_256(b"new-slice"),
+                        has_more: false,
                     },
                 }
                 .into_task_result(),
@@ -388,6 +389,7 @@ async fn test_ingest_process_several_runs() {
                     pull_result: PullResult::Updated {
                         old_head: Some(odf::Multihash::from_digest_sha3_256(b"new-slice")),
                         new_head: odf::Multihash::from_digest_sha3_256(b"new-slice-2"),
+                        has_more: false,
                     },
                 }
                 .into_task_result(),